backend/scubashop/requirements.txt
asgiref==3.7.2
boto3==1.35.5
botocore==1.35.5
certifi==2024.7.4
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==43.0.0
Django==5.0.6
django-allauth==0.63.6
django-filter==24.3
django-storages==1.14.4
djangorestframework==3.15.2
drf-yasg==1.21.7
idna==3.3
inflection==0.5.1
jmespath==1.0.1
packaging==21.3
paypalrestsdk==1.13.3
Pillow==9.0.0
pycparser==2.21
PyJWT==2.8.0
pyOpenSSL==24.2.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2021.3
# PyYAML==6.0.2
requests==2.32.3
s3transfer==0.10.2
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.10.0
uritemplate==4.1.1
urllib3==1.26.9
